refactor: update data types for ServiceResponseFileId to BIGINT in multiple tables, remove obsolete HttpExecutionDetailAuditsServiceAppHealthchecksLinks table, and add new HttpExecutionDetailAuditsServiceResponseFilesLinks table for execution tracking

+++ OrbitToolDatabase/dbo/Tables/DirectExecutionGroupServiceRequestFileLinkAudits.sql
@@ -12,7 +12,7 @@ CREATE TABLE [dbo].[DirectExecutionGroupServiceRequestFileLinkAudits] (
     [Id] BIGINT IDENTITY(1,1) NOT NULL,
     -- Public Identifier for UI/Secure Operations (GUID)
     [DirectExecutionGroupServiceRequestFileLinkId] INT NOT NULL,
-    [ServiceResponseFileId] INT NOT NULL,
+    [ServiceResponseFileId] BIGINT NOT NULL,
     -- Execution order within the audit
     [ExecutionOrder] INT NOT NULL CONSTRAINT DF_DirectExecutionGroupServiceRequestFileLinkAudits_ExecutionOrder DEFAULT 0,
     -- Timestamp when the execution started
